+++ CHANGELOG.md
@@ -16,6 +16,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## 2019-10-17
+
+### Changed [Added Nutanix/Hyper-V support to New-RubrikSnapshot]
+* Modified URI list in Get-RubrikAPIData for New-RubrikSnapshot to include support for Nutanix and Hyper-V
+* Addresses [Issue 466](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/466)
+
+## 2019-10-11
+
+### Changed [Added documentation to all private functions]
+* Added documentation in all private functions including synopsis, description, parameter help and examples where appropriate
+* Addresses [Issue 378](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/378)
+
 ## 2019-10-08
 
 ### Added [cmdlets to add and remove organizations]